fix(wb-scraper): card decode — weight/volume float (WB sends kg as fraction)
Live e2e against WB showed competitor_cards/prices/details/stocks decoded
to 0 rows: WB serves weight/volume as fractional kg (e.g. 0.04, 12.5), but
decode.go declared them int64. Go's strict json unmarshal then fails the
whole card body, dropping every card in the batch.

- types.go: CompetitorCard.Weight/Volume int64 -> float64
- decode.go: wbCardProduct.Weight/Volume int64 -> float64
- schemas: weight/volume REAL (SQLite) / DOUBLE PRECISION (PostgreSQL)

Also popup.js: default the "Коллектор" endpoint field unconditionally. A
`!= null` guard left it empty on first open (storage has no key yet), so
only the placeholder showed and COLLECT_START's `!endpoint` check silently
bailed — no tab opened on Start.

Verified live (1 search target): 200 positions + 2015 cards + 3515 prices
+ 55 details + 459 stocks in /tmp DB. vitrine_ads still 0 — separate ad
shape bug (array vs object), deferred.

+++ pkg/storage/postgres/wbscraper_schema.go
@@ -82,8 +82,8 @@ CREATE TABLE IF NOT EXISTS competitor_cards (
     rating        DOUBLE PRECISION NOT NULL DEFAULT 0,
     feedbacks     INTEGER NOT NULL DEFAULT 0,
     pics          INTEGER NOT NULL DEFAULT 0,
-    weight        BIGINT NOT NULL DEFAULT 0,
-    volume        BIGINT NOT NULL DEFAULT 0,
+    weight        DOUBLE PRECISION NOT NULL DEFAULT 0,
+    volume        DOUBLE PRECISION NOT NULL DEFAULT 0,
     colors        TEXT NOT NULL DEFAULT '',
     subject_id    BIGINT,
     panel_promo_id BIGINT
